Changes utils.get_keystone(token) to utils.check_is_admin(token)
We never use the client that get_keystone returns. We are actually
interested in whether the provided token belongs to an admin user.
This patch changes the function name to reflect this fact. It also
adds a check to make sure the token is from a user with the admin
role on the tenant provided in the admin_tenant_name configuration
option.

Change-Id: I657d14921e9a3db7fea4499144a77659bb3f390d
Closes-Bug: 1391866