Adding root_device_hint plugin
In some cases (eg. DRAC), the driver doesn't provide any information to
figure out the root device during RAID configuration. A workaround for
that is creating the RAID disks in 2 runs: one run for the root device
and another one for the rest. In between ironic-discoverd can be used to
examine the changes in the block devices.

This patch adds this capability as a plugin to gather the block device
serials, save it temporarly and expose the root device hint.

Change-Id: I71c42cb131a4e8f0b4fe4eb2f42ac7cde2bceea9