eDeploy: Store all of the facts collected by edeploy in Ironic DB
We want to store the benchmark data collected by the eDeploy
ramdisk. As this data is somewhat tricky to seperate out from the
facts collected by eDeploy, and the other facts may prove useful
for other cases, this patch just stores all of the collected data.
The collected facts are stored in the extra column of the Ironic
database under the 'edeploy_facts' key. This will allow us to do
manual analysis of the benchmark data before moving the node to
the available state.

Change-Id: Iadc5d1160cea4834184375adfba3bec2bf3d3d5c