Delay failure from pre-processing hooks
If something fails in the 'before_processing' hook, the only feedback
a user will receive is timeout. Actually, in many cases we're able
to find a node in cache and update it's state with reasonable error.

This patch collects exceptions from the pre-processing hooks and
raises them all after node is found in cache. If node is not found,
we assume that it might be due to hook failures.

Change-Id: I8f320504818b002981757dbc9bb1cf98b9829683
Closes-Bug: #1436276