Change the way error are returned from API to JSON
Previously we used just strings. It's not a breaking change in terms
of API, but it will lead to less readable errors for older clients.

Nice side effect of this change is that exceptions are no longer
reported as HTML.

In-tree client was updated, as separate one is not set up yet.

Change-Id: Ice3111e08f8716d904efc4869c6d2ded25e23e2e
Closes-Bug: #1465347