Stop using configuration options in top level code
This code makes assumption that importing oslo_config.cfg is enough
to get our options, which is of course not true. It happens to work now,
as we seem to import ironic_inspector.conf (actual module defining
configuration options) somewhere else before this code is run.

When building an RPM, however, we run tests a bit differently, and
this problems leads to NoSuchOptError: no such option:
introspection_delay_drivers.

Closes-Bug: #1494649
Change-Id: I3524222418a12270c78c6ee61d4a4afa1353306e