Ignore IPMI Address for IPMI Bridged nodes
When a node is behind an IPMI bridge the IPMI address stored in ironic
is not a trust worthy identifier for recognising a node returning its
introspection data, as it can be the same or multiple nodes. It will
also cause unique constraint errors as the database will not allow
multiple bmc_address attribute entries with the same value. This patch
adds detection for ipmi bridging to the get_ipmi_address function, so
if ipmi bridging is enabled then it returns None, and the node is added
without any bmc_address listed.

Change-Id: I09d17dffcde0f4a023bb03e1bce88bcf725cdf0a
Closes-Bug: #1488501