Convert eDeploy data so that rules can process it
eDeploy's return format is a list of lists, in order to use it with
introspection rules we must convert it to dictionary format. This patch
detects if the data is in eDeploy's format and will convert it storing
the new data in introspection_data['extra']. introspection_data['data']
is assumed unusable by introspection rules and will always be removed
from introspection_data after the data is stored in swift.

Change-Id: I25c7ef1f3d3122b80ea98e4d0e64ea20e2eec618
Closes-Bug: #1495649