[devstack] Use the coreos builder for the source build of the ramdisk
The source builder in DIB has been broken for a while now, and anyway
we have to use the CoreOS builder to create an IPA+inspector gate:
I5631866edee42d5358f692e4fbcd82f1a6a1d4db.

We can provide DIB as an option again as soon as it's fixed.

Change-Id: I9a53c3bedd4106a4b03647209ea2a49efa4b2a47