Allow specify log level for Error exception
For now, all Error exceptions logged with error level, but some exceptions
may be expected as correct flow and they shouldn't be logged as errors.

NotFoundInCacheError, for example, is raised when there isn't info in cache
about introspected node, this case may be handled by not_found_hook, so
this wouldn't be error anymore.

Change-Id: Ie537ccaef0035b2ef839c34fad0a5e6c9ba8f064