Add callback function to manage missing hooks
Before https://review.openstack.org/#/c/337043/,
stevedore.NamedExtensionManager returned KeyError
when calling with non-existing names and name_order=True.

After the mentioned change this is not longer true,
so we used the just added on_missing_entrypoints_callback option
to customize the behavior in this cases and make it raise
a custom exception MissingHookError.

Change-Id: I1f1edc0b7a82a16bf9be4113db61ee1cd0080db4
Closes-Bug: #1600141