LLC Hook: Fix patching Ironic ports
This patch corrects a problem with the LLC hook's use
of the node_info.patch_port function, ensuring the correct data is
passed to the function, not a unicode string.

This patch also ensures that the mac addresses for chassis-id and
port-id are passed to ironic in the expected Unix format. Ironic will
fail to validate the mac addresses without this.

Change-Id: Id5b729cd9e9beeb4f59ba6950162163d2fdf3a3a
Closes-Bug: #1629302
Closes-Bug: #1629303