LLC Hook: Do not assume interfaces are added to Ironic
The current local_link_connection hook assumes that all inspected
interfaces are either added to or already exist in ironic as ports. When
CONF.processing.add_ports is set to PXE, and a node has more interfaces
than the one that we PXE on, then there will be more interfaces in the
inventory than added to Ironic. This patch ensures we must handle that
scenario.

Change-Id: I42c98c571ae8cfc26f319b712b5ac5bad88bb75d
Closes-Bug: #1628991