Pass session directly to swiftclient
since v3.2.0 swiftclient supports instantiating the API client from
keystonauth session, and we already require this version as minimal in
ironic.

This patch removes existing workarounds in SwiftAPI code which were there
only to accomodate for absence of session handling in swiftclient,
and now the keystoneauth session built from the [swift] config section
is used directly to instantiate the swift API client.

Change-Id: I593156d122db62b49e2d383d3dbbbae691b7b904
Related-Bug: #1518938
Co-Authored-By: Pavlo Shchelokovskyy <shchelokovskyy@gmail.com>