grenade: Only 'enable_plugin ironic-inspector' if not already in conf
To support multi-node grenade jobs we need to move enabling of the
ironic-inspector devstack plugin out of the ironic-inspector grenade
settings file devstack/upgrade/settings. But if we add enabling of the
ironic-inspector devstack plugin to project-config it will cause the gate
to break as the plugin will be enabled twice.

This patch is similar to the change done in ironic for the same purpose.

This checks to see if the ironic-inspector devstack plugin has already been
enabled:
  * If it has already been enabled it will not enable it again.
  * If it has not yet been enabled it will enable it.

This will allow us to update project-config to enable the ironic-inspector
devstack plugin and not break the gate.

After all the changes have landed we will propose a follow-up patch to
remove all this additonal logic and the enabling of the ironic-inspector
devstack plugin.

Change-Id: Id1a5de9a8aefe751b63c1a6a9f2c942fe4c957b1
Closes-Bug: #1672448
(cherry picked from commit 69157e5554ef6ad10f3d6d3ffc313d4e498ec21b)