Allow periodics to terminate inspector
This patch allows a periodic task to shut down the inspector upon a
(critical) failure.  This is needed in case of a (future) PXE filter driver
malfunction to avoid inspector running with a broken filter.

The test_wsgi_service unit test module was refactored to address test
ordering dependency issues and to enhance the mocking.

Change-Id: I50e6408bf30416d8c7ff9e30be0e938cec12b7cd
Related-Bug: 1665666