Change example dnsmasq.conf in install doc for uefi
As described in https://bugs.launchpad.net/puppet-ironic/+bug/1714320,
a loop can occur in which dnsmasq always indicates ipxe.efi as the
bootfile if the client request has both UEFI and iPXE options set.
This fixes the documentation for the example dnsmasq.conf so that
loop won't occur.

In addition, some UEFI clients use architecture type 9 so that is
added to the example dnsmasq.conf.

Related-Bug: 1714320
Change-Id: Icdefba2b1be47622c9bd9c3bfb9b3413964c9dac