Properly init config in unit tests
upcoming policy-related changes need a properly initialized
global oslo_config's CONF instance, which inspector's unit tests
currently lack.

As a side effect of implementing this, the 'dbsync' module was
changed to not register it's CLI options right on its import.

Also, setting default log levels was moved to a function which
is already registered as the one providing defaults to options
in oslo_config's entrypoints. Config sample is updated accordingly.

Change-Id: I20bc537605062900d00fcc0343172774c1cd1363