Fix Py2/Py3 differences in write locking code
In Py2, a file object write method returns NoneType.

In Py3, a file object write method returns a count of
bytes written, which can be used to interpret success.

As this code, at least for now, still needs to work
on Py2 without raising what appears to be errors, we
need to remove the assumption that data is returned
upon write.

Should a write fail, IOError is still raised in Py2,
which means the existing exception handling should be
sufficent.

Also, added an explicit flush for before we release the file
lock, just to ensure that the data is actually written out of
python's buffers before the lock is released.

Change-Id: I1cae8f1cd2f7da39600d72a84fe041ff0a97e580
Closes-Bug: #1741035