Use the 'ironic' queue for the gate
When doing the 'gate' build have the ironic-inspector jobs be part of
the 'ironic' queue. More information on this is in [1]. The benefit of
this is that we are only impacted by ironic projects if we do this and
openstack/ironic is also using the 'ironic' queue.

From [1]:
Pipelines which use the dependent pipeline manager maintain separate
queues for groups of projects. When Zuul serializes a set of changes
which represent future potential project states, it must know about
all of the projects within Zuul which may have an effect on the
outcome of the jobs it runs. If project A uses project B as a library,
then Zuul must be told about that relationship so that it knows to
serialize changes to A and B together, so that it does not merge a
change to B while it is testing a change to A.

[1] https://docs.openstack.org/infra/zuul/feature/zuulv3/user/config.html

Change-Id: I735dc13c288bc16c9d4ea6bc0a63ca24a70b3272