Use processed bool as key in introspection_data DB table
When using store_data=database with sqlalchemy, running introspection results
in a DBDuplicateEntry error.  This happens because the query when adding an
entry uses both the primary key (uuid) and the processed flag, but the
processed flag is not a key.  This change makes it a key so that both
unprocessed and processed data can be stored in the table.

Note - since the previous migration hasn't been released yet this fixes it
without creating a new one.

Change-Id: I052594d529ae363fce50b1726169d86583bb1439
Story: #2004992
Task: #29463