Convert Swift connection errors into more obvious ones
Currently we handle Swift errors on putting/getting objects, but not
on connections. This patch improves it for better error messages.

This also means that commit 09938a397595b92f19f6415bb3cfec8be76fe6a3
is incomplete and does not actually fix the extra-hardware plugin
when Swift is absent. This change corrects it.

Change-Id: I0e8e76291653dd2e0cd3cde35d18a1e8f9c91853