Comment out race prone sqlite db model sync test
The nature of the unit tests with sqlite means that it is
easy to get into a race condition with the tests where one
worker thread makes changes to the database that essentially
causes another test to fail.

Since we don't run these sorts of sync tests in ironic for
similar reasons, I've commented out the test that is causing
numerous false failures due to race conditions.

Change-Id: Ia4927fd8eca671414bd77f3d0ef188ba94cc0e2d