Set default boot option to netboot for jobs using cirros
An upcoming ironic change changes the default boot option for nodes
to local booting as opposed to the long standing default netboot.

Some jobs still use cirros, which cannot be local booted from a
partition image as it is designed to write the ramdisk contents
to the partition upon the initial network boot.

Change-Id: I0d0c043de449048c3cc38ed7d7dac3e9e85974ef