Stop configuring install_command in tox.
Currently, we are overriding 'install_command' to use 'pip'. This is
considered poor behavior and 'python -m pip' should be used instead:

https://snarky.ca/why-you-should-use-python-m-pip/

It turns out that this is the the default value provided by tox:

https://tox.readthedocs.io/en/latest/config.html#conf-install_command

So we can remove the line and simply use the default value.

openstack-discuss thread
http://lists.openstack.org/pipermail/openstack-discuss/2020-April/014237.html

Bumping requirements and l-c with:
keystoneauth1 3.18.0
alembic 0.9.6
pika 0.10.0 (necessary for installation)
Change-Id: I849342bd077ab2f8bbda4e0a228c308eb7a4908b