Allow operators to disable clean_up sync
Some operators simply don't need the periodic clean up sync
task as part of their ironic deployment with ironic-inspector
as they are not using the PXE filtering to manage access for
machines as part of discovery.

In these "standalone" use cases, there is typically no neutron,
and everything is managed externally from ironic-inspector.

Change-Id: I0036b2cdb7d562e90855ccabd108392f2f97c6f9