Fix URL parsing to enable redfish_address matching
Fixes the ironic utilities logic to parse URLs in driver_info
properly.

Also adds ``redfish_address`` to the default list of fields to
evaluate for host matching.

Previously an operator could add the field to the list, but the
matching logic would not match the URL as it did not know to
decompose the url and identify the hostname portion of the url.

Resolves: rhbz#1670336
Story: 2008010
Task: 40660
Change-Id: Ice86e9ab3efb98b649141bdf7e1e2febdc9203a8