[DOC] Add note regarding ``[pxe_filter]deny_unknown_macs``
Add a note in the user guide section related to validate_interfaces
plug-in. When [pxe_filter]deny_unknown_macs is enabled no node
without a pre-enrolled port will boot into introspection, so one
cannot rely on the validate network interfaces plug-in to create
the PXE-booting port in this case.

This is a follow up to: https://review.opendev.org/753435

Change-Id: I41c7693f3f73ddb29725741075b9fb0b4c0ebd02