Fix database migrations and disable the non-standalone job
The pattern of adding a column and then reading a table with it
no longer works in SQLAlchemy 1.3.20. This has been reported
upstream [1] and apparently we're doing the wrong thing by reusing
a column object twice.

We need to disable the non-standalone job since it's really broken
now, and this fix is blocking bifrost.

[1] https://github.com/sqlalchemy/sqlalchemy/issues/5669

Change-Id: I2fb07413e8f421f39b24acf1272771ee2097b195