Re-enable ironic-inspector-tempest-managed-non-standalone job
The pymemcache lib was updated recently in upper-constraints to 3.4.0
and since then the ironic-inspector-tempest-managed-non-standalone job
has been broken.
There seems to be an issue between tooz memcache driver and pymemcache,
the change in upper-constraints has been reverted [1] and the failing
version blacklisted in global-requirements.
Re-enabling ironic-inspector-tempest-managed-non-standalone as it
will now pass correctly.

[1] https://review.opendev.org/760103

Change-Id: I4944860c265f9e5060c8ab276e7962a93a27da88