Use correct Node id attribute
According to the openstacksdk docs[1] the Node uuid is stored in the
id attribute. This change removes the get_node shim which adds the
uuid attribute, and replaces any calls to Node.uuid with Node.id. This
will stop the many log debug warnings about this attribute:

    DEBUG openstack.resource [-] Attribute [uuid] not found in
    [<openstack.resource._ComponentManager object at 0x7f417e43aa20>]:
    'uuid'. __getattribute__ /usr/lib/python3.6/site-packages/openstack/resource.py:623

Calls to NodeInfo.uuid or db.Node.uuid remain unchanged.

Change-Id: Icd3de82877c6a53d32b4c9fd3e500d3cd9d7fb17
Story: 2008379
Task: 41300