Remove grenade jobs from old stable branches
As discussed and approved during the ironic wallaby midcycle, the grenade
jobs in old stable branches, starting from EM branches and up the oldest
not-EM (at this moment train) are removed.

And remove lower-constraints job.

Also changing bifrost-integration-tinyipa-ubuntu-xenial to non-voting.
At the moment it's not working and we need to fix it in bifrost.
Bifrost CI in stein has some fundamental jobs not working properly
at the moment and it will require some effort to make it pass, so for
the sake of unblocking the work here in inspector, we go for an easy
temporary solution, also considering that xenial will be EOL in few months.

Change-Id: I7304593bf0e5cfa481761c1dc4740f4ad85cc0cf