Remove grenade jobs from old stable branches
As discussed and approved during the ironic wallaby midcycle, the grenade
jobs in old stable branches, starting from EM branches and up the oldest
not-EM (at this moment train) are removed.

Also remove lower-constraints job.

Change bandit version for testing to be inferior than 1.7.0 as Python
2.7 support was removed in that version.

Changing bifrost-integration-tinyipa-ubuntu-xenial to non-voting. At the
moment it's not working and we need to fix it in bifrost. Bifrost CI in
train has some fundamental jobs not working properly at the moment and
it will require some effort to make it pass, so for the sake of
unblocking the work here in inspector, we go for an easy temporary
solution, also considering that xenial will be EOL in few months.

Change-Id: If56029e8fd1bf164fb2d1272ae434b9ef1c4d19f