Fix node id vs uuid in processing_logger_prefix
this is a follow-up to Icd3de82877c6a53d32b4c9fd3e500d3cd9d7fb17

one more place was identified as trying to use uuid attr of a Node
object from openstacksdk instead of id attr.

In tests, use stricter mocking with spec_set (that actually fails to
access or set an atribute of a mock that is absent from the spec class
or instance) to guard against future possible changes.

Change-Id: I83c420d0e91e00f583a28833f4f710cf70b01fa8
Story: 2008379
Task: 41300